add page folders
