+++ src/Router.js
@@ -10,20 +10,20 @@ import UserContext from "./UserContext";
 import { NavigationBar } from "./components/NavigationBar";
 
 import Authenticator from "./Authenticator";
-import Home from "./Home";
+import Home from "./Homepage/Home";
 import { Upload } from "./Upload";
 import Watch from "./Watch";
 import Live from "./Live";
-import { Stream } from "./Stream";
-import Profile from "./Profile";
+import { Stream } from "./TribalPage/Stream";
+import Profile from "./SocialPage/Profile";
 import GetAccess from "./GetAccess";
 import Pending from "./Pending";
 import { MyLive } from "./MyLive";
 import BottomNavBar from "./components/bottomNavBar";
-import SearchQueries from "./searchQueries";
+import SearchQueries from "./SearchPage/searchQueries";
 import ViewProfile from "./ViewProfile";
 import MyList from "./MyList";
-import Settings, { WifiAndCellular } from "./Settings";
+import Settings, { WifiAndCellular } from "./SettingsPage/Settings";
 
 class PrivateRoute extends React.Component {
   // PrivateRoutes require authentication to access
