+++ src/SearchPage/searchQueries.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { API, graphqlOperation } from 'aws-amplify';
-import * as queries from './graphql/queries';
+import * as queries from '../graphql/queries';
 import { Link } from "react-router-dom";
 
 export default class SearchQueries extends React.Component {
